backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
langchain==0.1.0
openai==0.28.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0